input overhaul and a lot of refactoring and cleanup
handling the input had to change because of stdin blocking behaviour in
tokio. Now a task is spawned in the start of the program and a global
channel is used to communicate to it for getting input from stdin.
Some cleanup and refactoring for functions, especially in the net module.